HotFix Button to Delete a Role (Final Changes #1)
